koding hanya boleh koder

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\StudentController;
 use App\Http\Controllers\Cesemix\RanapMonitController;
 use App\Http\Controllers\RM\EklaimController;
+use App\Http\Middleware\CheckRole;
 
 Route::get('/', function () {
     return Inertia::render('Dashboard');
@@ -33,16 +34,16 @@ Route::prefix('casemix')->group(function () {
         Route::get('/list_billing_temp/{kode_reg}', [RanapMonitController::class, 'list_billing_temp'])->name('casemix.ranap-monit.list_billing_temp');
 
         Route::get('/list_procedure/{kode_reg}', [RanapMonitController::class, 'list_procedure'])->name('casemix.ranap-monit.list_procedure');
-        Route::post('/save_procedure', [RanapMonitController::class, 'save_procedure'])->name('casemix.ranap-monit.save_procedure');
-        Route::delete('/delete_procedure/{id}', [RanapMonitController::class, 'delete_procedure'])->name('casemix.ranap-monit.delete_procedure');
+        Route::post('/save_procedure', [RanapMonitController::class, 'save_procedure'])->middleware(['auth', CheckRole::class.':superadmin,koder'])->name('casemix.ranap-monit.save_procedure');
+        Route::delete('/delete_procedure/{id}', [RanapMonitController::class, 'delete_procedure'])->middleware(['auth', CheckRole::class.':superadmin,koder'])->name('casemix.ranap-monit.delete_procedure');
 
         Route::post('/update_monit_row/{kode_reg}', [RanapMonitController::class, 'update_monit_row'])->name('casemix.ranap-monit.update_monit_row');
 
         Route::get('/get_list_cppt/{kode_reg}', [RanapMonitController::class, 'get_list_cppt'])->name('casemix.pasien-inap.get_list_cppt');
 
         Route::get('/list_diagnosa/{kode_reg}', [RanapMonitController::class, 'list_diagnosa'])->name('casemix.ranap-monit.list_diagnosa');
-        Route::post('/save_diagnosa', [RanapMonitController::class, 'save_diagnosa'])->name('casemix.ranap-monit.save_diagnosa');
-        Route::delete('/delete_diagnosa/{id}', [RanapMonitController::class, 'delete_diagnosa'])->name('casemix.ranap-monit.delete_diagnosa');
+        Route::post('/save_diagnosa', [RanapMonitController::class, 'save_diagnosa'])->middleware(['auth', CheckRole::class.':superadmin,koder'])->name('casemix.ranap-monit.save_diagnosa');
+        Route::delete('/delete_diagnosa/{id}', [RanapMonitController::class, 'delete_diagnosa'])->middleware(['auth', CheckRole::class.':superadmin,koder'])->name('casemix.ranap-monit.delete_diagnosa');
 
         Route::get('/download-xls', [RanapMonitController::class, 'download_pasien_data'])->name('casemix.ranap-monit.download_pasien_data_xls');
     });
